Implement aspect-ratio interpolation support.
Based on the discussion in https://github.com/w3c/csswg-drafts/issues/4953
aspect ratios should be interpolated by the log of their value preserving
the same visual speed whether they are wide or narrow.

This patch adds a new interpolation type for aspect ratios and uses it
for interpolating the CSS aspect-ratio property.

Bug:1156160
Change-Id: I6cbff0abef54290de559a7625e24e3c4e1f1e0e9