usb: Add a manual test for USB transfer completion order
This manual test attempts to reproduce a case seen in issue 1153647
where parallel calls to transferIn(), recommended for performance, can
return data out of order.

This test is a starting point in the investigation as it only sends a
small amount of data while the original report involves much larger
transfers.

Bug: 1153647
Change-Id: I4f4b8f5ad502d08ae7c4fb9cc22889047029afc4