Hyphenate the last word if only one word in the line
Blink (legacy and NG) and WebKit does not hyphenate the last
word in a paragraph, except when it is a single-word paragraph.
However, when the containing block is narrow, sometimes
authors want a long last word to be hyphenated.

This patch changtes the heursitic to allow hyphenation if the
line is a single-word line.

CSS Text 4 has a property to control this behavior, but no
browsers have implemented it yet.

Bug: 1022415
Change-Id: Ibd8d8b132e540a142f18efc09a37fed2b207bfdf